[Chore] #212 - Auth 네이밍 수정
