+++ Smeem-iOS/Smeem-iOS/Presentation/AuthBottomSheet/Login/ViewModel/LoginViewModel.swift
@@ -82,7 +82,7 @@ final class LoginViewModel: ViewModel {
             }
             .flatMap { type -> AnyPublisher<Void, Never> in
                 return Future<Void, Never> { promise in
-                    AuthAPI.shared.loginAPI(param: LoginRequest(social: type,
+                    AuthService.shared.loginAPI(param: LoginRequest(social: type,
                                                                 fcmToken: UserDefaultsManager.fcmToken)) { result in
                         switch result {
                         case .success(let response):
